Implements API#buildFile
buildFile assumes the request body is a JSON array of string point ids. For
each point specified in this array, it adds its point file entry to a string,
which is Cached for a brief time using a simple hash as a cache key.

This hash is returned in the response body for use in downloading the points
file.